+++ CHANGELOG.md
@@ -11,6 +11,8 @@ These usually have no immediately visible impact on regular users
 ### Added
 
 -   Is public toggle for annotations
+-   Undo/Redo
+    -   shape movement
 
 ### Fixes